Made a new package for equation generator
