+++ Phase0/src/EquationBuilder/AddBedmasEquationBuilder.java
@@ -1,0 +1,16 @@
+package EquationBuilder;
+
+public class AddBedmasEquationBuilder extends BedmasEquationBuilder{
+    public void buildOperator(){
+        bedmasEquation.setOperator();
+    }
+    public void buildAnswer(){
+
+    }
+    public void buildOperand1(){
+
+    }
+    public void buildOperand2(){
+
+    }
+}
