--- Phase0/src/MultiplyBedmasEquationBuilder.java
@@ -1,14 +1,0 @@
-public class MultiplyBedmasEquationBuilder extends BedmasEquationBuilder{
-    public void buildOperator(){
-
-    }
-    public void buildAnswer(){
-
-    }
-    public void buildOperand1(){
-
-    }
-    public void buildOperand2(){
-
-    }
-}
