+++ Phase0/test/BinaryExpressionTreeTest.java
@@ -1,0 +1,2 @@
+public class BinaryExpressionTreeTest {
+}
